Add User Permissions, Error Handling, Display Users + Group/Roles
Added the handling what happens when a user does not have the permissions necessary to perform a task. Added an asking Entity Framework for a list of users and roles, we are going to combine those objects into one list of objects and send it to the front-end, where we are going to display it in an administrative view.